+++ src/ext/java/org/opentripplanner/ext/reportapi/resource/ReportResource.java
@@ -26,7 +26,7 @@ public class ReportResource {
   @SuppressWarnings("unused")
   public ReportResource(@Context OTPServer server) {
     this.transferService = server.getRouter().transitModel.getTransferService();
-    this.index = server.getRouter().transitModel.index;
+    this.index = server.getRouter().transitModel.getTransitModelIndex();
   }
 
   @GET