+++ src/ext/java/org/opentripplanner/ext/transferanalyzer/DirectTransferAnalyzer.java
@@ -55,7 +55,7 @@ public class DirectTransferAnalyzer implements GraphBuilderModule {
     DataImportIssueStore issueStore
   ) {
     /* Initialize transit index which is needed by the nearby stop finder. */
-    transitModel.index = new TransitModelIndex(transitModel);
+    transitModel.setTransitModelIndex(new TransitModelIndex(transitModel));
     transitModel.getStopModel().index();
 
     LOG.info("Analyzing transfers (this can be time consuming)...");