+++ src/test/java/org/opentripplanner/model/impl/OtpTransitServiceBuilderLimitPeriodTest.java
@@ -9,7 +9,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.opentripplanner.graph_builder.DataImportIssueStore;
 import org.opentripplanner.model.Direction;
 import org.opentripplanner.model.PickDrop;
 import org.opentripplanner.model.StopPattern;
@@ -114,7 +113,7 @@ public class OtpTransitServiceBuilderLimitPeriodTest {
     assertEquals(1, patternInT2.getScheduledTimetable().getTripTimes().size());
 
     // Limit service to last half of month
-    subject.limitServiceDays(new ServiceDateInterval(D2, D3), new DataImportIssueStore(false));
+    subject.limitServiceDays(new ServiceDateInterval(D2, D3));
 
     // Verify calendar
     List<ServiceCalendar> calendars = subject.getCalendars();